# Optimized MiddlewareMixin coroutine check.