Cache `get_response`’s coroutine status. Afaiu `get_response` is never expected to change.

Repeating the check on every is unnecessary work - about 1 microsecond each time, on my machine's Python 3.10:

```
In [2]: async def foo(): return

In [3]: def bar(): return

In [4]: %timeit asyncio.iscoroutinefunction(foo)
979 ns ± 16.9 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)

In [5]: %timeit asyncio.iscoroutinefunction(bar)
1.04 µs ± 10.5 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```