@kezabelle 2021-12-20
Hey Adam,
I'd spotted this a while back too, but I didn't know enough about how the machinery of async and ASGI works to say for certain that the wrapping up of callables into a get_response would always be _static_ after init. So I'm glad you're pushing it, where I was too uncertain :)

When I was looking at it, I was thinking about re-using the `_is_coroutine` attribute which is (unfortunately) conditionally set, and doing `getattr(self, '_is_coroutine', False) is True`; presumably it's slightly faster the way you've done it by adding another attribute?