@adamchainz 2021-12-20
>  I'd spotted this a while back too, but I didn't know enough about how the machinery of async and ASGI works to say for certain that the wrapping up of callables into a get_response would always be static after init. So I'm glad you're pushing it, where I was too uncertain :)

Great minds think alike!

> When I was looking at it, I was thinking about re-using the `_is_coroutine` attribute which is (unfortunately) conditionally set, and doing `getattr(self, '_is_coroutine', False)`; presumably it's slightly faster the way you've done it by adding another attribute?

Absolutely right, I missed that, and have updated to use it.

> This is where we find folks dynamically munging middleware functions. 😜 (I agree that's not expected, so the change should be OK.)

This would already be broken, as `async_check()` is not rerun